Expose the comparable, stata, complete xlsx and codebook files
A lot of new files can be downloaded... in the future, this
prepares their download but the files themselves cannot be
downloaded yet as the survey results have not been published yet.

But they will be there soon. You won't even have noticed it via
the git log because this won't be published until the survey
results are.