Update text on download page and adjust text padding
Text updates for the different options on the download page. The
custom reports option is changed the most as the download button
and the years are moved to the left column (instead of being above
the preview in the right column).

Left column text for custom reports is padded by 10px to align with
heading.