Add countries that are not present in previous reports to timeline
The timeline goes through older reports to find previous data and
then adds survey data up to that point. However countries that were
added in the most recent survey will then not be found. We need to
check if a country was found in previous data and if not then we
explicitly add it for the most recent year.