Remove Comparable questions from download data user interface
The comparable questions title is not what IBP wants it to be so
it should be removed until a decision about it is made instead of
using a name that can be interpreted incorrectly.

This should be reverted once a decision is made and the title and
text updated according to the decision.