Link dots to available drive directories.
Linking directly to gdrive documents within their year folders is
problematic given we haven't got a definitive fiscal year property from
the API. Instead we'll now link to the top level report type directory
for the country in question. e.g. Argentina/Year-End Report.

Dots will be linked if a directory for the report type exists on the
gdrive.

Related to #113.