Add Sao Tome and Comoros to world map javascript
- Add these two islands and simple cirles
- Remove previous marker for Sao Tome and related special handling

Sao Tome was previously added to the map as a marker and had special
handling in map.js to display the colour and label correctly. This
commit adds the custom countries directly to the map, so no special
handling is necessary, and the countries scale correctly as different
zoom levels. Although they are still represented as simple circles, this
is still a better solution than using markers.